app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["lucas <development@waetzig.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
streamlit = ">=1.48.0,<2.0.0"
pandas = ">=2.3.1,<3.0.0"
sqlalchemy = ">=2.0.42,<3.0.0"
psycopg2-binary = "^2.9.10"
transformers = "^4.55.2"
torch = "^2.8.0"
plotly = "^6.3.0"
nbformat = "^5.10.4"
gensim = "^4.3.3"
nltk = "^3.9.1"
numpy = ">=1.26.0,<2.0.0"
scipy = ">=1.10.0,<1.13.0"
bertopic = "^0.17.3"
sentence-transformers = "^5.1.0"
hdbscan = "^0.8.40"
umap-learn = "^0.5.9.post2"
matplotlib = "^3.10.5"
spacy = "^3.8.7"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
ipykernel = "^6.30.1"
isort = "^6.0.1"
python-dotenv = "^1.1.1"
watchdog = "^6.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
